requirements.txt
Django>=3.1.7
django-filter>=2.4.0
mysqlclient>=2.0.3
djangorestframework==3.12.3
drf-yasg>=1.20.0
Celery>=5.0.5
redis>=3.5.3
django-redis>=4.12.1
drf-renderer-xlsx
pandas
django-celery-beat

channels>=3.0.3
django-cors-headers>=3.7.0
django-debug-toolbar>=3.2.1